requirements.txt
requests
python-dotenv
feedparser
yfinance
akshare
